Update pages-and-routes.md
Added a description for displaying local images. The conventions of how to reference local images are unclear especially for beginners. I lost 15 minutes because I somehow managed to screw it and had to resort to discord to find your answer to Chris Wells Wood on 20.06.2023 :)

I know it is a small thing and very obvious in hindsight